+++ ext_emconf.php
@@ -10,7 +10,7 @@ $EM_CONF[$_EXTKEY] = [
     'constraints' => [
         'depends' => [
             'typo3' => '10.4.11-10.4.99',
-            'jobrouter_base' => '0.1.0-0.1.99',
+            'jobrouter_base' => '0.1.1-0.1.99',
             'jobrouter_connector' => '0.12.0-0.12.99',
         ],
         'conflicts' => [],