+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Dashboard widget "JobData Transmission Errors"
+
 ## [0.12.2] - 2021-03-02
 
 ### Changed